первая загрузка

+++ app/src/main/java/com/example/notesapp/data/repository/NotesRepository.kt
@@ -1,9 +1,11 @@
 package com.example.notesapp.data.repository
 
 import android.content.Context
+import com.example.notesapp.R
 import com.example.notesapp.data.apiservice.ApiService
 import com.example.notesapp.data.database.dao.NotesDao
 import com.example.notesapp.data.database.entitys.Notes
+import com.example.notesapp.utils.ConnectReceiver
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.*
@@ -12,6 +14,7 @@ import kotlinx.coroutines.launch
 class NotesRepository(
     private val notesDao: NotesDao,
     private val apiService: ApiService,
+    private val connectReceiver: ConnectReceiver,
     private val applicationContext: Context
 ) {
 
@@ -22,6 +25,9 @@ class NotesRepository(
 
     private val serviceError = MutableStateFlow("")
     val serviceErrorFlow: StateFlow<String> = serviceError.asStateFlow()
+
+    private val isLoaded = MutableStateFlow(false)
+    val isLoadedFlow: StateFlow<Boolean> = isLoaded.asStateFlow()
 
     fun setInsertedOrEditedIdNull() {
         insertedOrEditedId = 0L
@@ -40,12 +46,24 @@ class NotesRepository(
             emit(notesDao.getNoteById(noteId).firstOrNull())
         }
 
+    fun loadRemoutData() {
+        isLoaded.value = true
+
+
+
+        isLoaded.value = false
+    }
+
     fun addNote(note: Notes) {
         CoroutineScope(Dispatchers.Default).launch {
             try {
-                val resultId: Long = apiService.addNote(note)
-                insertedOrEditedId = resultId
-                isNoteEdited.value = true
+                if(connectReceiver.isConnectStatusFlow.value) {
+                    val resultId: Long = apiService.addNote(note)
+                    insertedOrEditedId = resultId
+                    isNoteEdited.value = true
+                } else {
+                    serviceError.value = applicationContext.getString(R.string.operation_not_possible)
+                }
             } catch (e: Exception) {
                 serviceError.value = e.message.toString()
             }
@@ -55,9 +73,12 @@ class NotesRepository(
     fun deleteNote(note: Notes) {
         CoroutineScope(Dispatchers.Default).launch {
             try {
-                apiService.deleteNote(note)
-                isNoteEdited.value = true
-                //isNoteDeleted.value = true
+                if(connectReceiver.isConnectStatusFlow.value) {
+                    apiService.deleteNote(note)
+                    isNoteEdited.value = true
+                } else {
+                    serviceError.value = applicationContext.getString(R.string.operation_not_possible)
+                }
             } catch (e: Exception) {
                 serviceError.value = e.message.toString()
             }
